+++ geopackage-sdk/src/main/java/mil/nga/geopackage/tiles/overlay/FeatureRowData.java
@@ -5,6 +5,7 @@ import java.util.Map;
 
 import mil.nga.geopackage.geom.GeoPackageGeometryData;
 import mil.nga.sf.Geometry;
+import mil.nga.sf.GeometryEnvelope;
 import mil.nga.sf.GeometryType;
 import mil.nga.sf.geojson.FeatureConverter;
 
@@ -22,9 +23,24 @@ public class FeatureRowData {
     private Map<String, Object> values;
 
     /**
+     * Id column
+     */
+    private String idColumn;
+
+    /**
      * Geometry column
      */
     private String geometryColumn;
+
+    /**
+     * Constructor
+     *
+     * @param values column names and values
+     * @since 6.3.1
+     */
+    public FeatureRowData(Map<String, Object> values) {
+        this(values, null);
+    }
 
     /**
      * Constructor
@@ -33,7 +49,20 @@ public class FeatureRowData {
      * @param geometryColumn geometry column name
      */
     public FeatureRowData(Map<String, Object> values, String geometryColumn) {
+        this(values, null, geometryColumn);
+    }
+
+    /**
+     * Constructor
+     *
+     * @param values         column names and values
+     * @param idColumn       id column name
+     * @param geometryColumn geometry column name
+     * @since 6.3.1
+     */
+    public FeatureRowData(Map<String, Object> values, String idColumn, String geometryColumn) {
         this.values = values;
+        this.idColumn = idColumn;
         this.geometryColumn = geometryColumn;
     }
 
@@ -44,6 +73,30 @@ public class FeatureRowData {
      */
     public Map<String, Object> getValues() {
         return values;
+    }
+
+    /**
+     * Get the id column name
+     *
+     * @return id column
+     * @since 6.3.1
+     */
+    public String getIdColumn() {
+        return idColumn;
+    }
+
+    /**
+     * Get the id
+     *
+     * @return id
+     * @since 6.3.1
+     */
+    public Long getId() {
+        Long id = null;
+        if (idColumn != null) {
+            id = (Long) values.get(idColumn);
+        }
+        return id;
     }
 
     /**
@@ -61,7 +114,11 @@ public class FeatureRowData {
      * @return geometry data
      */
     public GeoPackageGeometryData getGeometryData() {
-        return (GeoPackageGeometryData) values.get(geometryColumn);
+        GeoPackageGeometryData geometryData = null;
+        if (geometryColumn != null) {
+            geometryData = (GeoPackageGeometryData) values.get(geometryColumn);
+        }
+        return geometryData;
     }
 
     /**
@@ -70,7 +127,42 @@ public class FeatureRowData {
      * @return geometry
      */
     public Geometry getGeometry() {
-        return getGeometryData().getGeometry();
+        Geometry geometry = null;
+        GeoPackageGeometryData geometryData = getGeometryData();
+        if (geometryData != null) {
+            geometry = geometryData.getGeometry();
+        }
+        return geometry;
+    }
+
+    /**
+     * Get the geometry type
+     *
+     * @return geometry type
+     * @since 6.3.1
+     */
+    public GeometryType getGeometryType() {
+        GeometryType geometryType = null;
+        Geometry geometry = getGeometry();
+        if (geometry != null) {
+            geometryType = geometry.getGeometryType();
+        }
+        return geometryType;
+    }
+
+    /**
+     * Get the geometry envelope
+     *
+     * @return geometry envelope
+     * @since 6.3.1
+     */
+    public GeometryEnvelope getGeometryEnvelope() {
+        GeometryEnvelope envelope = null;
+        GeoPackageGeometryData geometryData = getGeometryData();
+        if (geometryData != null) {
+            envelope = geometryData.getOrBuildEnvelope();
+        }
+        return envelope;
     }
 
     /**